Move schema package to metricbeat
The schema package is also used in libbeat which created a dependency on metricbeat. To make the schema package also reusable in other beats an remove the dependency, the package was moved to libbeat.
