+++ libbeat/common/schema/error.go
