+++ libbeat/common/schema/error_test.go
