+++ libbeat/common/schema/errors.go
