+++ libbeat/common/schema/errors_test.go
