+++ libbeat/common/schema/mapstriface/mapstriface_test.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/elastic/beats/libbeat/common"
-	s "github.com/elastic/beats/metricbeat/schema"
+	s "github.com/elastic/beats/libbeat/common/schema"
 	"github.com/stretchr/testify/assert"
 )
