+++ libbeat/common/schema/mapstrstr/mapstrstr_test.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"github.com/elastic/beats/libbeat/common"
-	s "github.com/elastic/beats/metricbeat/schema"
+	s "github.com/elastic/beats/libbeat/common/schema"
 	"github.com/stretchr/testify/assert"
 )
