+++ libbeat/common/schema/schema_test.go
