+++ libbeat/processors/add_cloud_metadata/provider_aws_ec2.go
@@ -2,8 +2,8 @@ package add_cloud_metadata
 
 import (
 	"github.com/elastic/beats/libbeat/common"
-	s "github.com/elastic/beats/metricbeat/schema"
-	c "github.com/elastic/beats/metricbeat/schema/mapstriface"
+	s "github.com/elastic/beats/libbeat/common/schema"
+	c "github.com/elastic/beats/libbeat/common/schema/mapstriface"
 )
 
 // AWS EC2 Metadata Service
