+++ metricbeat/module/postgresql/bgwriter/data.go
@@ -3,8 +3,8 @@ package bgwriter
 import (
 	"time"
 
-	s "github.com/elastic/beats/metricbeat/schema"
-	c "github.com/elastic/beats/metricbeat/schema/mapstrstr"
+	s "github.com/elastic/beats/libbeat/common/schema"
+	c "github.com/elastic/beats/libbeat/common/schema/mapstrstr"
 )
 
 var schema = s.Schema{
